+++ src/lib.rs
@@ -6,23 +6,22 @@ extern crate reqwest;
 extern crate serde;
 #[macro_use]
 extern crate serde_derive;
+#[macro_use]
 extern crate serde_json;
 extern crate tempdir;
 extern crate url;
 extern crate uuid;
 
 /*
-extern crate base64;
 extern crate hyper;
-#[macro_use(mime, __mime__ident_or_ext)]
 */
 
 mod attachment;
 mod client;
 /*
 mod design;
-mod document;
 */
+mod document;
 mod error;
 mod revision;
 /*
@@ -32,17 +31,13 @@ mod view;
 pub mod action;
 */
 pub mod path;
-/*
 pub mod testing;
-*/
-
 pub use attachment::{Attachment, SavedAttachment, UnsavedAttachment};
 pub use client::IntoUrl;
 /*
 pub use design::{Design, DesignBuilder, ViewFunction};
-pub use document::Document;
 */
-
+pub use document::Document;
 pub use error::{Error, ErrorResponse};
 pub use path::{AttachmentName, AttachmentPath, DatabaseName, DatabasePath, DesignDocumentName, DesignDocumentPath,
                DocumentId, DocumentPath, IntoAttachmentPath, IntoDatabasePath, IntoDesignDocumentPath,