+++ src/testing/mod.rs
@@ -4,4 +4,6 @@ mod fake_server;
 
 pub use self::fake_server::FakeServer;
 pub use document::DocumentBuilder;
+/*
 pub use view::{IsGrouped, IsReduced, IsUnreduced, ViewResponseBuilder};
+*/